change amount to big decimal

+++ server/src/main/java/com/app/bicycle/entities/Payment.java
@@ -18,7 +18,7 @@ public class Payment extends BaseEntity {
     private User user;
 
     @Column
-    private Double amount;
+    private BigDecimal amount;
 
     @Column
     private Date date;
@@ -31,11 +31,11 @@ public class Payment extends BaseEntity {
         this.user = user;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
